+++ src/main/java/br/com/henrique/service/EstadoService.java
@@ -1,13 +1,16 @@
 package br.com.henrique.service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import br.com.henrique.dto.EstadoDto;
 import br.com.henrique.model.Estado;
 import br.com.henrique.repository.EstadoRepository;
+import br.com.henrique.service.exception.ObjectFoundException;
 import br.com.henrique.service.exception.ObjectNotFoundException;
 
 @Service
@@ -18,11 +21,15 @@ public class EstadoService {
     
     // Lista Estado
     public List<Estado> findAll() {
-        List<Estado> estados = new ArrayList<Estado>();
-        estados = repositEstado.findAll();
+        List<Estado> estados = repositEstado.findAll();
         
         return estados;
     }    
+    
+    // Lista Estados com Paginação
+    public Page<Estado> findAllPage(Pageable pageable) {
+        return repositEstado.findAll(pageable);
+    }
  
     // Busca por Estado
     public Estado findById(String sigla) {
@@ -33,23 +40,33 @@ public class EstadoService {
         return estado;
     }
     
-    // Inclui Empresa
-    public Estado addEstado(Estado estado) {
-        return repositEstado.save(estado);
+    // Inclui Empresa - DTO
+    public Estado addEstado(EstadoDto estadoDto) {
+        Estado estadoBuscaID = repositEstado.findById(estadoDto.getSigla()).orElse(null);
+        if (estadoBuscaID != null) {
+            throw new ObjectFoundException("Estado já cadastrado !");
+        }            	
+        return repositEstado.save(estadoDto.converteToEntity());
     }
     
-    // Altera Estado
-    public void updateEstado(String sigla, Estado estado) {
+    // Altera Estado - DTO
+    public void updateEstado(String sigla, EstadoDto estadoDto) {
         Estado estadoAtualizado = this.findById(sigla);
-        estadoAtualizado.setNome(estado.getNome());
+        if (estadoAtualizado == null) {
+            throw new ObjectNotFoundException("Estado nao encontrado !");
+        }                
+        estadoAtualizado.setNome(estadoDto.getNome());
         
         repositEstado.save(estadoAtualizado);
     }
     
     // Excluir Estado
     public void deletaEstado(String sigla) {
-        this.findById(sigla);
-
+        Estado estadoExcluir = this.findById(sigla);
+        if (estadoExcluir == null) {
+            throw new ObjectNotFoundException("Estado nao encontrado !");
+        }     
+        
         repositEstado.deleteById(sigla);
     }
 }
